+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # CHANGELOG.md
 
-## 6.2.1.3 (2022-11-10)
+## 6.3.0.0 (2022-11-14)
+
+Features:
+  - Integrated navigator search in slashcommands (Only when Next Experience is anabled)
+  - Option to local store last slashcommands per instance, scroll through them with arrow up, down key
 
 Fixes / changes:
   - Improve copycells for cells with long text content (Issue #318)
   - Fix scriptsync icons click event inside forms in studio (Issue #319)
   - Fix slushbucket when filtered to not loose eventhandlers (issue #322)
+  - Removed hist / nav / fav slashcommands
+  - Improved way of handling slashcommands when opening page in current window
+  - Removed space to complete slashcommand, to not interfere with new serach function (tab still works)
 
 
 ## 6.2.1.0 (2022-11-03)